Migrate to test_file.py.
